+++ clevertap-hms/src/main/java/com/clevertap/android/hms/HmsSdkHandler.java
@@ -8,9 +8,12 @@ import static com.clevertap.android.sdk.pushnotification.PushConstants.LOG_TAG;
 import android.content.Context;
 import android.text.TextUtils;
 import com.clevertap.android.sdk.CleverTapInstanceConfig;
-import com.huawei.agconnect.config.AGConnectServicesConfig;
+import com.huawei.agconnect.AGConnectOptions;
+import com.huawei.agconnect.AGConnectOptionsBuilder;
 import com.huawei.hms.aaid.HmsInstanceId;
 import com.huawei.hms.api.HuaweiApiAvailability;
+
+import org.jetbrains.annotations.TestOnly;
 
 /**
  * Implementation of {@link IHmsSdkHandler}
@@ -21,16 +24,27 @@ class HmsSdkHandler implements IHmsSdkHandler {
 
     private final CleverTapInstanceConfig mConfig;
 
+    private final AGConnectOptions options ;
+
     HmsSdkHandler(final Context context, final CleverTapInstanceConfig config) {
         this.context = context.getApplicationContext();
         mConfig = config;
+        options = new AGConnectOptionsBuilder().build(context);
     }
+
+    @TestOnly
+    HmsSdkHandler(final Context context, final CleverTapInstanceConfig config,final AGConnectOptions op) {
+        this.context = context.getApplicationContext();
+        mConfig = config;
+        options = op;
+    }
+
 
     @Override
     public String appId() {
         String appId = null;
         try {
-            appId = AGConnectServicesConfig.fromContext(context).getString(APP_ID_KEY);
+            appId = options.getString(APP_ID_KEY);
         } catch (Throwable t) {
             mConfig.log(LOG_TAG, HMS_LOG_TAG + "HMS availability check failed.");
         }