+++ clevertap-xps/src/main/java/com/clevertap/android/xps/CTXiaomiMessageHandler.java
@@ -19,6 +19,7 @@ import com.clevertap.android.sdk.interfaces.INotificationParser;
 import com.clevertap.android.sdk.interfaces.IPushAmpHandler;
 import com.clevertap.android.sdk.pushnotification.PushNotificationHandler;
 import com.clevertap.android.sdk.pushnotification.fcm.IFcmMessageHandler;
+import com.xiaomi.channel.commonutils.android.Region;
 import com.xiaomi.mipush.sdk.ErrorCode;
 import com.xiaomi.mipush.sdk.MiPushClient;
 import com.xiaomi.mipush.sdk.MiPushCommandMessage;
@@ -91,8 +92,12 @@ public class CTXiaomiMessageHandler implements IMiMessageHandler, IPushAmpHandle
                 Logger.d(LOG_TAG, "onReceiveRegisterResult() : Token is null or empty");
                 return INVALID_TOKEN;
             }
-            PushNotificationHandler.getPushNotificationHandler().onNewToken(context, token, XPS
-                    .getType());
+            String region = MiPushClient.getAppRegion(context);
+            region =TextUtils.isEmpty(region)? Region.Global.name() : region;
+            Logger.v("default CTXiaomiMessageHandler: onReceiveRegisterResult | MiPushClient.getAppRegion(context) returns region="+region);
+            XPS.setServerRegion(region);
+            PushNotificationHandler.getPushNotificationHandler().onNewToken(context, token, XPS.getType());
+
             return TOKEN_SUCCESS;
         } catch (Throwable t) {
             Logger.d(LOG_TAG, "onReceiveRegisterResult() : Exception: ", t);