+++ clevertap-xps/src/test/java/com/clevertap/android/xps/XpsTestConstants.kt
@@ -6,5 +6,9 @@ internal interface XpsTestConstants {
         const val MI_TOKEN = "abcdefghij"
         const val MI_APP_KEY = "1234"
         const val MI_APP_ID = "xyzwr"
+        const val MI_REGION_IN = "in1"
+        const val MI_REGION_EU = "eu1"
+        const val MI_REGION_SG = "sg1"
+        const val MI_REGION_US = "us1"
     }
 }